refactor: rename dao to store
